StackUsingArrayUpdated.cpp
Now we make some changes to this new version. As we know that in earlier case we are bounded with the total size of array, But now if the size of array is full but still we want to push another element to our array we can do that by creating a new array of double the size of previous array.